fix(editor): force dynamic /editor, refactor load logic, use named Tiptap export; chore(footer): correct starter attribution
