+++ app/tag-data.json
@@ -1,37 +1,83 @@
 {
-  "vercel": 15,
-  "hypeburner": 1,
   "zed": 1,
   "deltadb": 1,
   "github": 3,
   "sequoia": 1,
   "funding": 1,
-  "growthbook": 2,
-  "paddle": 1,
-  "mcp": 1,
-  "cloudinary": 2,
-  "mux": 2,
-  "zitadel": 1,
-  "stytch": 3,
-  "imgix": 1,
-  "algolia": 2,
-  "netlify": 6,
-  "workos": 1,
-  "planetscale": 6,
-  "better-auth": 1,
-  "timescale": 2,
-  "knock": 2,
-  "liveblocks": 1,
+  "bun": 1,
+  "firebase": 1,
+  "heroku": 4,
   "neon": 10,
-  "polar": 1,
-  "browserbase": 1,
-  "inngest": 1,
+  "netlify": 6,
+  "planetscale": 6,
+  "pulumi": 2,
+  "sitecore": 1,
+  "stellate": 1,
+  "vercel": 14,
+  "apollo": 2,
+  "directus": 3,
+  "fusionauth": 1,
+  "gitlab": 2,
+  "nile": 1,
+  "remix": 1,
+  "sanity": 3,
   "supabase": 7,
+  "vite": 2,
+  "angular": 1,
+  "cypress": 1,
+  "grafbase": 6,
+  "hashicorp": 1,
+  "tabnine": 1,
+  "terraform": 1,
+  "xata": 5,
+  "aws-amplify": 1,
+  "jira": 1,
+  "kontent-ai": 1,
+  "docker": 1,
+  "wordpress": 3,
+  "launchdarkly": 6,
+  "posthog": 3,
+  "resend": 1,
+  "statsig": 3,
+  "render": 4,
+  "upstash": 1,
+  "wundergraph": 1,
+  "builderio": 1,
+  "datadog": 2,
+  "new-relic": 1,
+  "strapi": 2,
+  "astro": 1,
+  "ubicloud": 1,
+  "cloudflare": 2,
+  "langchain": 3,
+  "prisma": 2,
+  "weaviate": 1,
+  "deno": 1,
+  "nextjs": 2,
+  "pinecone": 2,
+  "contentful": 2,
+  "openai": 2,
+  "tembo": 1,
+  "timescale": 2,
+  "anthropic": 1,
+  "devtools": 2,
   "newsletter": 6,
   "stripe": 3,
-  "privy": 1,
-  "cursor": 1,
+  "knock": 2,
+  "autoblocks": 1,
+  "convex": 1,
+  "eppo": 1,
+  "bucket": 1,
+  "figma": 1,
+  "windsurf": 1,
+  "chroma": 1,
   "databricks": 2,
+  "mastra": 1,
+  "stytch": 3,
+  "revenuecat": 1,
+  "algolia": 2,
+  "mux": 2,
+  "frontegg": 1,
   "plaid": 2,
   "descope": 1,
   "ory": 1,
@@ -39,65 +85,18 @@
   "sentry": 1,
   "shadcn": 1,
   "postman": 1,
-  "launchdarkly": 6,
-  "sanity": 3,
-  "revenuecat": 1,
-  "frontegg": 1,
-  "xata": 5,
-  "mastra": 1,
-  "figma": 1,
-  "openai": 2,
-  "windsurf": 1,
-  "chroma": 1,
-  "statsig": 3,
-  "langchain": 3,
-  "datadog": 2,
-  "eppo": 1,
-  "bucket": 1,
-  "autoblocks": 1,
-  "convex": 1,
-  "devtools": 2,
-  "pulumi": 2,
-  "contentful": 2,
-  "prisma": 2,
-  "posthog": 3,
-  "pinecone": 2,
-  "render": 4,
-  "anthropic": 1,
-  "nextjs": 2,
-  "heroku": 4,
-  "tembo": 1,
-  "gitlab": 2,
-  "cloudflare": 2,
-  "deno": 1,
-  "strapi": 2,
-  "weaviate": 1,
-  "wordpress": 3,
-  "astro": 1,
-  "directus": 3,
-  "ubicloud": 1,
-  "grafbase": 6,
-  "builderio": 1,
-  "new-relic": 1,
-  "apollo": 2,
-  "upstash": 1,
-  "wundergraph": 1,
-  "resend": 1,
-  "docker": 1,
-  "kontent-ai": 1,
-  "aws-amplify": 1,
-  "jira": 1,
-  "vite": 2,
-  "angular": 1,
-  "cypress": 1,
-  "hashicorp": 1,
-  "tabnine": 1,
-  "terraform": 1,
-  "fusionauth": 1,
-  "nile": 1,
-  "remix": 1,
-  "bun": 1,
-  "firebase": 1,
-  "sitecore": 1,
-  "stellate": 1
+  "cloudinary": 2,
+  "privy": 1,
+  "cursor": 1,
+  "liveblocks": 1,
+  "polar": 1,
+  "browserbase": 1,
+  "inngest": 1,
+  "growthbook": 2,
+  "workos": 1,
+  "better-auth": 1,
+  "paddle": 1,
+  "mcp": 1,
+  "zitadel": 1,
+  "imgix": 1
 }
