+++ components/Footer.tsx
@@ -26,7 +26,7 @@ export default function Footer() {
         </div>
         <div className="mb-8 text-sm text-gray-500 dark:text-gray-400">
           <Link href="https://github.com/timlrx/tailwind-nextjs-starter-blog">
-            Made with ❤️ in North Carolina
+            Built on Tailwind Next.js Starter Blog
           </Link>
         </div>
       </div>
